housing-consumer/pyproject.toml
[project]
name = "housing-consumer"
version = "0.1.0"
description = ""
authors = [
    {name = "adam",email = "adamezouhairi@outlook.fr"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "confluent-kafka (>=2.8.2,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
